[MOD]opd-4878 修改最近1年营业总额、最近2年营业总额由原先的Long类型修改为BigDecimal

+++ src/main/java/org/srm/source/cux/domain/entity/RcwlSupplierHeader.java
@@ -100,10 +100,10 @@ public class RcwlSupplierHeader extends ExpandDomain {
     private Long years;
     @ApiModelProperty(value = "近1年营业额（万元）")
     @NotNull
-    private Long oneProfit;
+    private BigDecimal oneProfit;
     @ApiModelProperty(value = "近2年营业额（万元）")
     @NotNull
-    private Long twoProfit;
+    private BigDecimal twoProfit;
     @ApiModelProperty(value = "启用标识", required = true)
     @NotNull
     private Integer enabledFlag;
@@ -331,22 +331,22 @@ public class RcwlSupplierHeader extends ExpandDomain {
     /**
      * @return 近1年营业额（万元）
      */
-    public Long getOneProfit() {
+    public BigDecimal getOneProfit() {
         return oneProfit;
     }
 
-    public void setOneProfit(Long oneProfit) {
+    public void setOneProfit(BigDecimal oneProfit) {
         this.oneProfit = oneProfit;
     }
 
     /**
      * @return 近2年营业额（万元）
      */
-    public Long getTwoProfit() {
+    public BigDecimal getTwoProfit() {
         return twoProfit;
     }
 
-    public void setTwoProfit(Long twoProfit) {
+    public void setTwoProfit(BigDecimal twoProfit) {
         this.twoProfit = twoProfit;
     }
